SetList
en esta oportunidad cambie las listas contenidas por setter al igual que el idFestival ahora es un entero no un long como en versiones anteriores

+++ src/test/TestAgregarFestival.java
@@ -11,7 +11,7 @@ public class TestAgregarFestival {
 FestivalABM abm = new FestivalABM();
 		
 		try {
-			long ultimoIdFestival;
+			int ultimoIdFestival;
 			ultimoIdFestival = abm.agregar("Nombre","temporada" , LocalDate.now(), LocalDate.now(),
 					1111.111,2222.222,3333.333,4444.444,5555.555);
 			System.out.printf("Id cliente: %d", ultimoIdFestival);
